requirements.txt
tensorflow-macos==2.13.0
tensorflow-metal==1.0.1
mediapipe==0.10.14
opencv-python==4.10.0.84
numpy==1.26.4
scikit-learn==1.4.2
matplotlib==3.8.4
